+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ops/autotests/Red_Auto.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ops.Auto_Tests;
+package org.firstinspires.ftc.teamcode.ops.autotests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -37,9 +37,9 @@ import org.firstinspires.ftc.teamcode.bots.TestBot;
 import org.firstinspires.ftc.teamcode.components.DriveTrain;
 
 
-@Autonomous(name="Crab_Test_Auto", group="Auto_Tests")
+@Autonomous(name="Red_Auto", group="autotests")
 //@Disabled
-public class Crab_Test_Auto extends LinearOpMode {
+public class Red_Auto extends LinearOpMode {
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
@@ -79,30 +79,22 @@ public class Crab_Test_Auto extends LinearOpMode {
            robot.driveTrain.encoderDrive(1, -10);
            robot.driveTrain.encoderDrive(1, 10); */
 
-           //robot.driveTrain.crabLeft(0.15);
-       /*    robot.driveTrain.encoderCrab(1, 0.5, 0);
-           robot.driveTrain.gyroRotate(1, 0.5, true, false);
+           robot.driveTrain.crabEncoderLeft(0.5, 3);
+           robot.driveTrain.gyroRotate(-5, 0.2, true, false);
            robot.driveTrain.pause(0.5);
-           robot.driveTrain.moveForward(0.8, -1);
+           robot.driveTrain.moveForward(0.85, -1);
            robot.driveTrain.gyroRotate(90, 0.5, true, false);
-           robot.driveTrain.moveForward(0.5, 0.5);
+           robot.driveTrain.moveForward(0.7, 0.5);
            robot.driveTrain.pause(1);
            robot.grapple.servoMoveDown();
            robot.grapple.servo2MoveDown();
            robot.driveTrain.pause(2);
            robot.driveTrain.moveBackward(1, 0.5);
            //robot.driveTrain.move(.15,-1,1);
-           robot.driveTrain.gyroRotate(90, 0.7, true, false);
+           robot.driveTrain.gyroRotate(90, 0.5, true, false);
            robot.grapple.servoMoveUp();
            robot.grapple.servo2MoveUp();
-           robot.driveTrain.moveForward(0.8, 0.5); */
-
-/*
-        runtime.reset();
-        while(runtime.seconds() < 2 && opModeIsActive()) {
-            robot.driveTrain.mechanumMotorDiagonalPower(0.25, -0.25);
-        }
-*/
+           robot.driveTrain.moveForward(0.8, 0.5);
