+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ops/game/Game_TeleOP.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ops.TeleOP_Tests;
+package org.firstinspires.ftc.teamcode.ops.game;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -42,8 +42,8 @@ import org.firstinspires.ftc.teamcode.components.Intake;
 import org.firstinspires.ftc.teamcode.components.WebCamera;
 
 
-@TeleOp(name="testTeleOP1", group="template")
-public class Template_TeleOP_test1 extends LinearOpMode {
+@TeleOp(name="Game_TeleOP", group="game")
+public class Game_TeleOP extends LinearOpMode {
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
     private TestBot robot = null;
