+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ops/game/Plan_B_Autonomous_Red_Side.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.ops.gyrotest;
+package org.firstinspires.ftc.teamcode.ops.game;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -8,9 +8,9 @@ import org.firstinspires.ftc.teamcode.bots.TestBot;
 import org.firstinspires.ftc.teamcode.components.DriveTrain;
 
 
-@Autonomous(name="Time_Auto_Red_Platform", group="gyrotest")
+@Autonomous(name="Plan_B_Autonomous_Red_Side", group="gyrotest")
 //@Disabled
-public class Time_Auto_Red_Platform extends LinearOpMode {
+public class Plan_B_Autonomous_Red_Side extends LinearOpMode {
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
@@ -50,47 +50,8 @@ public class Time_Auto_Red_Platform extends LinearOpMode {
            robot.driveTrain.encoderDrive(1, -10);
            robot.driveTrain.encoderDrive(1, 10); */
 
-        //Move the robot down the field twoards the build platform
-        robot.driveTrain.moveBackward(.95, -.75);
 
-        //Rotate the back of the robot twoards the build platform
-        robot.driveTrain.gyroRotate(-85, .75, true, false);
-
-        //Move the robot so that it is touching the build platform
-        robot.driveTrain.moveBackward(0.7, -0.5);
-
-        //Pause to let the robot stop moving
-        robot.driveTrain.pause(2 );
-
-        //Move the servos down to grapple the build platform
-        //robot.grapple.servoMoveDown();
-        robot.grapple.servo2MoveDown();
-
-        //Pause
-        robot.driveTrain.pause(2);
-
-        //Pull the platform twoards the build zone
-        robot.driveTrain.moveForward(1.25, -0.5);
-
-        //Move because the robot can not fine adjust to make the gyro happy with the platform in tow
-        //robot.driveTrain.move(1, -1, 1);
-
-        robot.driveTrain.gyroRotate(95, .5);
-
-        //Move the servos up to release the platform
-        //robot.grapple.servoMoveUp();
-        //robot.grapple.servo2MoveUp();
-
-        //Push the build platform to the wall to score it
-        robot.driveTrain.moveBackward(.5, .75);
-
-        //Pause to let the robot stop moving
-        robot.driveTrain.pause(.25);
-
-        //Move the robot to park under the skybridge
-        robot.driveTrain.moveForward(0.9,  .75);
-
-        robot.driveTrain.pause(5 );
+        robot.driveTrain.moveBackward(.5, -.5);
         /*
         runtime.reset();
