+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ops/game/Time_Auto_Blue_No_Platform.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.ops.gyrotest;
+package org.firstinspires.ftc.teamcode.ops.game;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -8,7 +8,7 @@ import org.firstinspires.ftc.teamcode.bots.TestBot;
 import org.firstinspires.ftc.teamcode.components.DriveTrain;
 
 
-@Autonomous(name="Time_Auto_Blue_No_Platform", group="gyrotest")
+@Autonomous(name="Time_Auto_Blue_No_Platform", group="game")
 //@Disabled
 public class Time_Auto_Blue_No_Platform extends LinearOpMode {
