+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ops/game/Time_Auto_Red_No_Platform.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.ops.gyrotest;
+package org.firstinspires.ftc.teamcode.ops.game;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -8,9 +8,9 @@ import org.firstinspires.ftc.teamcode.bots.TestBot;
 import org.firstinspires.ftc.teamcode.components.DriveTrain;
 
 
-@Autonomous(name="Plan_B_Autonomous_Red_Side", group="gyrotest")
+@Autonomous(name="Time_Auto_Red_No_Platform", group="game")
 //@Disabled
-public class Plan_B_Autonomous_Red_Side extends LinearOpMode {
+public class Time_Auto_Red_No_Platform extends LinearOpMode {
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
@@ -41,29 +41,54 @@ public class Plan_B_Autonomous_Red_Side extends LinearOpMode {
         runtime.reset();
 
         double angle1 = robot.gyroNavigator.getAngle();
-        //   double angle2 = robot.gyroNavigator2.getAngle();
 
-         /*  robot.driveTrain.encoderDrive(1, -0.5);
-           robot.driveTrain.gyroRotate(90, 0.5, true, false);
-           robot.driveTrain.encoderDrive(1, -40);
-           robot.driveTrain.gyroRotate(-90, 0.5, true, false);
-           robot.driveTrain.encoderDrive(1, -10);
-           robot.driveTrain.encoderDrive(1, 10); */
+        //SKYSTONE ANALYSIS AND PICKUP PROCEDURE
 
+        //Move the robot down the field towards the build platform
+        robot.driveTrain.moveBackward(.30, .75);
 
-        robot.driveTrain.moveBackward(.5, -.5);
-        /*
+        //Rotate the back of the robot towards the build platform
+        robot.driveTrain.gyroRotate(-95, .75, true, false);
+
+        //Move the robot so that it is touching the build platform
+        robot.driveTrain.moveBackward(0.6, 0.5);
+
+        //Pause to let the robot stop moving
+        robot.driveTrain.pause(.25);
+
+        //Move the servos down to grapple the build platform
+        robot.ramp.ramp2Down();
+        robot.ramp.rampDown();
+
+        //Pause to let stone fall out
+        robot.driveTrain.pause(.25);
+
+        //Move ramp back up
+        robot.ramp.ramp2Up();
+        robot.ramp.rampUp();
+
+        //Move the robot forward so that it is next to the parking spot
+        robot.driveTrain.moveForward(0.6, 0.5);
+
+        //Turn the robot so that it can drive under the bridge
+        robot.driveTrain.gyroRotate(85, .5, true, false);
+
+        //Park under the SkyBridge
+        robot.driveTrain.moveForward(.25, 0.5);
+
+        robot.driveTrain.pause(.25);
+
         runtime.reset();
 
         while(runtime.seconds() < 60){
             robot.driveTrain.moveForward(0.5, 1);
             if(robot.skystoneFinder.canSeeSkystone()){
-                robot.driveTrain.moveForward(.5, .25);
+                robot.driveTrain.move(.5, .1,  .1);
 
             }
         }
 
-    //With Phone Camera Mounted on Side Pannel
+        //With Phone Camera Mounted on Side Pannel
 
         //robot.logger.logInfo("runOpMode", "Angles: 1:%f", angle1);
 
@@ -76,13 +101,13 @@ public class Plan_B_Autonomous_Red_Side extends LinearOpMode {
         // rotate 45 degrees away from bridge
         robot.driveTrain.gyroRotate(-45, .5);
         // stop do vision
-            //Get x-y-z coordinates from vuforia of skystone
+        //Get x-y-z coordinates from vuforia of skystone
 
-            //If no coordinates, move to set position
+        //If no coordinates, move to set position
 
-            //If corrdinate is found (found x and y position), then find the x-y-z position of skystone
+        //If corrdinate is found (found x and y position), then find the x-y-z position of skystone
 
-            //If x - y matches (23< x <25)recorded position for skystone
+        //If x - y matches (23< x <25)recorded position for skystone
 
 
 
@@ -97,7 +122,7 @@ public class Plan_B_Autonomous_Red_Side extends LinearOpMode {
         // after block pickup, return to set position which is tile adjacent to
         //   alliance and neutral bridge
 
-*/
+
         // Show the elapsed game time.
         robot.logger.logInfo("runOpMode", "===== [ Autonomous Complete ] Run Time: %s", runtime.toString());
         telemetry.update();
